+++ pkg/identityserver/application_access.go
@@ -27,6 +27,7 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/identityserver/store"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+	"go.thethings.network/lorawan-stack/v3/pkg/unique"
 )
 
 var (
@@ -227,6 +228,8 @@ func (is *IdentityServer) getApplicationCollaborator(ctx context.Context, req *t
 	return res, nil
 }
 
+var errApplicationNeedsCollaborator = errors.DefineFailedPrecondition("application_needs_collaborator", "every application needs at least one collaborator with all rights")
+
 func (is *IdentityServer) setApplicationCollaborator(ctx context.Context, req *ttnpb.SetApplicationCollaboratorRequest) (*pbtypes.Empty, error) {
 	// Require that caller has rights to manage collaborators.
 	if err := rights.RequireApplication(ctx, req.ApplicationIdentifiers, ttnpb.RIGHT_APPLICATION_SETTINGS_COLLABORATORS); err != nil {
@@ -235,25 +238,50 @@ func (is *IdentityServer) setApplicationCollaborator(ctx context.Context, req *t
 
 	err := is.withDatabase(ctx, func(db *gorm.DB) error {
 		store := is.getMembershipStore(ctx, db)
+		existingRights, err := store.GetMember(
+			ctx,
+			&req.Collaborator.OrganizationOrUserIdentifiers,
+			req.ApplicationIdentifiers.GetEntityIdentifiers(),
+		)
+		if err != nil && !errors.IsNotFound(err) {
+			return err
+		}
+		existingRights = existingRights.Implied()
+		newRights := ttnpb.RightsFrom(req.Collaborator.Rights...).Implied()
+		addedRights := newRights.Sub(existingRights)
+		removedRights := existingRights.Sub(newRights)
 
-		if len(req.Collaborator.Rights) > 0 {
-			newRights := ttnpb.RightsFrom(req.Collaborator.Rights...)
-			existingRights, err := store.GetMember(
-				ctx,
-				&req.Collaborator.OrganizationOrUserIdentifiers,
-				req.ApplicationIdentifiers.GetEntityIdentifiers(),
-			)
-
-			if err != nil && !errors.IsNotFound(err) {
+		// Require the caller to have all added rights.
+		if len(addedRights.GetRights()) > 0 {
+			if err := rights.RequireApplication(ctx, req.ApplicationIdentifiers, addedRights.GetRights()...); err != nil {
 				return err
 			}
-			// Require the caller to have all added rights.
-			if err := rights.RequireApplication(ctx, req.ApplicationIdentifiers, newRights.Sub(existingRights).GetRights()...); err != nil {
+		}
+
+		// Unless we're deleting the collaborator, require the caller to have all removed rights.
+		if len(newRights.GetRights()) > 0 && len(removedRights.GetRights()) > 0 {
+			if err := rights.RequireApplication(ctx, req.ApplicationIdentifiers, removedRights.GetRights()...); err != nil {
 				return err
 			}
-			// Require the caller to have all removed rights.
-			if err := rights.RequireApplication(ctx, req.ApplicationIdentifiers, existingRights.Sub(newRights).GetRights()...); err != nil {
+		}
+
+		if removedRights.IncludesAll(ttnpb.RIGHT_APPLICATION_ALL) {
+			memberRights, err := is.getMembershipStore(ctx, db).FindMembers(ctx, req.ApplicationIdentifiers.GetEntityIdentifiers())
+			if err != nil {
 				return err
+			}
+			var hasOtherOwner bool
+			for member, rights := range memberRights {
+				if unique.ID(ctx, member) == unique.ID(ctx, &req.Collaborator.OrganizationOrUserIdentifiers) {
+					continue
+				}
+				if rights.Implied().IncludesAll(ttnpb.RIGHT_APPLICATION_ALL) {
+					hasOtherOwner = true
+					break
+				}
+			}
+			if !hasOtherOwner {
+				return errApplicationNeedsCollaborator.New()
 			}
 		}